backend/requirements.txt
google-cloud-storage
google-cloud-vision
google-cloud-logging
Flask~=2.0
Flask-Cors>=3.0
